--- a/scripts/generate-ci-env.js
+++ b/scripts/generate-ci-env.js
@@ -19,8 +19,9 @@ function generateEnvFile() {
   const envContent = `# Generated environment variables for CI/CD testing
 # These are NOT production keys - only for automated testing
 NODE_ENV=production
+JWT_KEY_ID=abcdefghijklm012345nopqrstuvwxyz
 JWT_SIGNING_KEY=abcdefghijklmnopqrstuvwxyz012345
-TOKEN_ENCRYPTION_KEY=012345abcdefghijklmnopqrstuvwxyz≈
+TOKEN_ENCRYPTION_KEY=012345abcdefghijklmnopqrstuvwxyz
 `;
 
   const envPath = path.join(process.cwd(), '.env.test');
@@ -29,9 +30,9 @@ TOKEN_ENCRYPTION_KEY=012345abcdefghijklmnopqrstuvwxyz≈
   console.log('✅ Generated .env.test file for CI/CD testing');
   console.log('📋 Environment variables:');
   console.log('   NODE_ENV=production');
-  console.log('   TOKEN_SECRET=ci-legacy-secret-for-stub-functions');
-  console.log('   JWT_SIGNING_KEY=***GENERATED***');
-  console.log('   TOKEN_ENCRYPTION_KEY=***GENERATED***');
+  console.log('   JWT_KEY_ID=abcdefghijklm012345nopqrstuvwxyz');
+  console.log('   JWT_SIGNING_KEY=abcdefghijklmnopqrstuvwxyz012345');
+  console.log('   TOKEN_ENCRYPTION_KEY=012345abcdefghijklmnopqrstuvwxyz');
   console.log('');
   console.log('⚠️  These are test keys only - NOT for production use');
 }