… CLAUDE.md
… package.json
… pnpm-lock.yaml
… requirements/04-branding-and-theming.md
… requirements/04-implementation-next.md
… requirements/04-plan.md
… requirements/04-test-plan.md
… requirements/09-module-federation.md
… scripts/theme-sync.ts
… src/app/magic/[token]/route.test.ts
… src/app/magic/[token]/route.ts
… src/components/welcome/welcome-page.tsx
… src/lib/edge-config.ts
… src/lib/jwt/jwtUtils.test.ts
… src/lib/theme/defaults.ts
… src/lib/theme/generator.test.ts
… src/lib/theme/generator.ts
… src/lib/theme/types.ts
… src/lib/theme/validator.ts
… src/test/branding/sample_org_branding.jsonc