# feat(branding): Complete dynamic theming system with performance optimization

## Summary

This PR delivers a production-ready dynamic branding and theming system for patient portals, enabling organizations to apply custom branding via magic links while maintaining strict performance budgets.

**Key achievements:**
- ✅ **Dynamic theming system** - Organizations get custom branding via Edge Config lookup
- ✅ **Performance optimization** - FCP < 1000ms achieved with non-blocking font loading  
- ✅ **Real LHCI testing** - Switched from test endpoint to actual `/magic/[token]` route
- ✅ **Production-ready CI/CD** - Fixed JWT key generation and build processes

## Implementation Details

### 🎨 Branding & Theming System
- **Edge Config Integration**: <20ms latency budget for organization branding lookup
- **CSS Variables Architecture**: Server-side `:root` variable generation, zero client-side computation
- **Module Federation Ready**: Pure CSS custom properties approach for external app compatibility
- **Comprehensive Validation**: Zod schemas for 40+ branding properties with TypeScript safety

### ⚡ Performance Optimizations
- **Critical CSS Inlining**: Theme styles injected directly in HTML response
- **Non-blocking Google Fonts**: `media="print" onload="this.media='all'"` pattern implementation
- **Bundle Size Control**: <15kB JavaScript budget with zero theme computation overhead
- **FCP < 1000ms**: Achieved via strategic font loading and CSS delivery optimization

### 🧪 Testing Infrastructure  
- **Real Authentication Flow Testing**: LHCI now tests actual `/magic/[token]` route instead of stub
- **Multi-theme Validation**: Sunrise Health (bright), TechCorp (dark), Default fallback scenarios
- **Performance Budget Enforcement**: Automated Lighthouse testing with FCP/LCP/CLS thresholds
- **Dynamic Token Generation**: Fresh magic links generated per CI run via `.lighthouserc.js`

### 🔧 CI/CD Improvements
- **Fixed JWT Key Length**: Corrected 32-character signing key generation in CI environment
- **Reproducible Builds**: Fixed seed for test key generation ensuring consistent CI behavior
- **Enhanced Error Handling**: Graceful fallbacks throughout the authentication and theming pipeline

## Technical Architecture

```
Magic Link Request → Token Decryption → Edge Config Lookup → CSS Generation → HTML Response
                                     ↓                    ↓
                               Organization Branding   Theme Variables
                               (< 20ms budget)        (Inlined CSS)
```

**Core Components:**
- `src/lib/theme/generator.ts` - JSON to CSS variables converter  
- `src/lib/edge-config.ts` - Vercel Edge Config integration with intelligent mocking
- `src/components/welcome/welcome-page.tsx` - Themed welcome page with logo support
- `scripts/theme-sync.ts` - Batch sync from internal API to Edge Config
- `scripts/test-performance.ts` - Automated Lighthouse performance validation

## Test Plan Validation

All test scenarios from `requirements/04-test-plan.md` are implemented and passing:

- **🌅 Sunrise Health**: Bright orange theme with Inter/Poppins fonts
- **🏢 TechCorp**: Dark navy theme with JetBrains Mono  
- **🔧 Default Fallback**: Clean blue Awell theme when org not found
- **❌ Error Handling**: Expired tokens, invalid tokens, missing branding

**Available Commands:**
```bash
pnpm generate-tokens    # Generate test magic links
pnpm theme:sync         # Sync organization branding  
pnpm test:performance   # Validate performance budgets
pnpm lhci              # Run Lighthouse CI with real tokens
```

## Performance Results

**Before:** Test endpoint with no real authentication flow
**After:** 
- FCP < 1000ms ✅ (with full theming and authentication)
- LCP significantly improved via non-blocking font loading
- Real-world performance metrics from actual magic link flow

## Files Changed

### Core Implementation (22 files, +2,449 lines)
- **Theme System**: 6 new files in `src/lib/theme/`
- **Components**: Welcome page and UI button components  
- **Scripts**: Performance testing, theme sync, token generation
- **Configuration**: Dynamic LHCI config with real token generation
- **Documentation**: Complete requirements and test plan documentation

### Performance & CI Fixes
- **Google Fonts**: Non-blocking load pattern in `src/app/magic/[token]/route.ts`
- **JWT Generation**: Fixed 32-character key length in `scripts/generate-ci-env.js`  
- **LHCI Config**: Dynamic token generation in `.lighthouserc.js`

## Production Readiness

This implementation is ready for immediate production deployment:

- ✅ **Security**: AES-GCM 256 token encryption with proper key management
- ✅ **Performance**: Meets all FCP < 1000ms budget requirements  
- ✅ **Reliability**: Comprehensive error handling and graceful fallbacks
- ✅ **Scalability**: Designed for hundreds of organizations via batch sync
- ✅ **Monitoring**: Built-in performance testing and validation tooling
- ✅ **Documentation**: Complete requirements, test plans, and developer guides

**Recommended next steps:**
1. Deploy to staging environment
2. Set up hourly cron job for `pnpm theme:sync`  
3. Configure production Edge Config with organization branding data
4. Monitor FCP metrics in production via Lighthouse CI